Major Optimizations

+++ tpms/waveshare.h
@@ -6,32 +6,32 @@
 // Custom Headers
 #include "config.h"
 
-/* LCD Size */
+// LCD Size
 #define LCD_WIDTH   320
 #define LCD_HEIGHT  240
 
-/* Colors */
-#define WHITE               0xFFFF
-#define BLACK               0x0000    
-#define BLUE                0x001F  
-#define BRED                0xF81F
-#define GRED                0xFFE0
-#define GBLUE               0x07FF
-#define RED                 0xF800
-#define MAGENTA             0xF81F
-#define GREEN               0x07E0
-#define CYAN                0x7FFF
-#define YELLOW              0xFFE0
-#define BROWN               0xBC40 
-#define BRRED               0xFC07 
-#define GRAY                0x8430 
-#define DARKBLUE            0x01CF  
-#define LIGHTBLUE           0x7D7C   
-#define GRAYBLUE            0x5458 
-#define LIGHTGREEN          0x841F 
-#define LGRAY               0xC618 
-#define LGRAYBLUE           0xA651
-#define LBBLUE              0x2B12 
+// Colors
+#define WHITE        0xFFFF
+#define BLACK        0x0000    
+#define BLUE         0x001F  
+#define BRED         0xF81F
+#define GRED         0xFFE0
+#define GBLUE        0x07FF
+#define RED          0xF800
+#define MAGENTA      0xF81F
+#define GREEN        0x07E0
+#define CYAN         0x7FFF
+#define YELLOW       0xFFE0
+#define BROWN        0xBC40 
+#define BRRED        0xFC07 
+#define GRAY         0x8430 
+#define DARKBLUE     0x01CF  
+#define LIGHTBLUE    0x7D7C   
+#define GRAYBLUE     0x5458 
+#define LIGHTGREEN   0x841F 
+#define LGRAY        0xC618 
+#define LGRAYBLUE    0xA651
+#define LBBLUE       0x2B12 
 
 void LCD_Write_Command(uint8_t data);
 
@@ -51,4 +51,4 @@ void drawString(uint16_t x, uint16_t y, const char * pString, sFONT* Font, uint1
 
 void initDisplay(void);
 
-void displayTPMS( byte grid_num, double tire_pressure, double tire_temperature );
+void displayTPMS( byte grid_num, float tire_pressure, float tire_temperature );
